refactor: move chart_state to types/xy_chart/store
