+++ src/components/chart.tsx
@@ -2,7 +2,7 @@ import classNames from 'classnames';
 import { Provider } from 'mobx-react';
 import React, { CSSProperties, Fragment } from 'react';
 import { SpecsParser } from '../specs/specs_parser';
-import { ChartStore } from '../state/chart_state';
+import { ChartStore } from '../types/xy_chart/store/chart_state';
 import { htmlIdGenerator } from '../utils/commons';
 import { AnnotationTooltip } from './annotation_tooltips';
 import { ChartResizer } from './chart_resizer';
