+++ src/components/highlighter.tsx
@@ -1,7 +1,7 @@
 import { inject, observer } from 'mobx-react';
 import React from 'react';
 import { isPointGeometry } from '../types/xy_chart/rendering/rendering';
-import { ChartStore } from '../state/chart_state';
+import { ChartStore } from '../types/xy_chart/store/chart_state';
 
 interface HighlighterProps {
   chartStore?: ChartStore;
