+++ src/components/tooltips.tsx
@@ -2,7 +2,7 @@ import classNames from 'classnames';
 import { inject, observer } from 'mobx-react';
 import React from 'react';
 import { TooltipValue, TooltipValueFormatter } from '../types/xy_chart/utils/interactions';
-import { ChartStore } from '../state/chart_state';
+import { ChartStore } from '../types/xy_chart/store/chart_state';
 
 interface TooltipProps {
   chartStore?: ChartStore;
